// src/api/back/models/StoreCategoryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StoreCategoryDto
 */
export interface StoreCategoryDto {
    /**
     * 
     * @type {string}
     * @memberof StoreCategoryDto
     */
    category: string;
}

export function StoreCategoryDtoFromJSON(json: any): StoreCategoryDto {
    return StoreCategoryDtoFromJSONTyped(json, false);
}

export function StoreCategoryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StoreCategoryDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'category': json['category'],
    };
}

export function StoreCategoryDtoToJSON(value?: StoreCategoryDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'category': value.category,
    };
}