// src/api/back/models/StoreCreateProductDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StoreCreateProductDto
 */
export interface StoreCreateProductDto {
    /**
     * 
     * @type {string}
     * @memberof StoreCreateProductDto
     */
    category: string;
    /**
     * 
     * @type {string}
     * @memberof StoreCreateProductDto
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof StoreCreateProductDto
     */
    imageKey: string;
    /**
     * 
     * @type {string}
     * @memberof StoreCreateProductDto
     */
    itemId: string;
    /**
     * 
     * @type {number}
     * @memberof StoreCreateProductDto
     */
    price: number;
}

export function StoreCreateProductDtoFromJSON(json: any): StoreCreateProductDto {
    return StoreCreateProductDtoFromJSONTyped(json, false);
}

export function StoreCreateProductDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StoreCreateProductDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'category': json['category'],
        'title': json['title'],
        'imageKey': json['imageKey'],
        'itemId': json['itemId'],
        'price': json['price'],
    };
}

export function StoreCreateProductDtoToJSON(value?: StoreCreateProductDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'category': value.category,
        'title': value.title,
        'imageKey': value.imageKey,
        'itemId': value.itemId,
        'price': value.price,
    };
}