// src/api/back/models/StoreItemHolderDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    StoreHatItemDto,
    StoreHatItemDtoFromJSON,
    StoreHatItemDtoFromJSONTyped,
    StoreHatItemDtoToJSON,
} from './';

/**
 * 
 * @export
 * @interface StoreItemHolderDto
 */
export interface StoreItemHolderDto {
    /**
     * 
     * @type {string}
     * @memberof StoreItemHolderDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof StoreItemHolderDto
     */
    title: string;
    /**
     * 
     * @type {StoreHatItemDto}
     * @memberof StoreItemHolderDto
     */
    hat?: StoreHatItemDto;
}

export function StoreItemHolderDtoFromJSON(json: any): StoreItemHolderDto {
    return StoreItemHolderDtoFromJSONTyped(json, false);
}

export function StoreItemHolderDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StoreItemHolderDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'],
        'hat': !exists(json, 'hat') ? undefined : StoreHatItemDtoFromJSON(json['hat']),
    };
}

export function StoreItemHolderDtoToJSON(value?: StoreItemHolderDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'title': value.title,
        'hat': StoreHatItemDtoToJSON(value.hat),
    };
}