// src/api/back/models/UploadedImageDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UploadedImageDto
 */
export interface UploadedImageDto {
    /**
     * 
     * @type {string}
     * @memberof UploadedImageDto
     */
    url: string;
    /**
     * 
     * @type {string}
     * @memberof UploadedImageDto
     */
    key: string;
}

export function UploadedImageDtoFromJSON(json: any): UploadedImageDto {
    return UploadedImageDtoFromJSONTyped(json, false);
}

export function UploadedImageDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UploadedImageDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'url': json['url'],
        'key': json['key'],
    };
}

export function UploadedImageDtoToJSON(value?: UploadedImageDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'url': value.url,
        'key': value.key,
    };
}